remove path comments from .css files
Before this commit, the build process added comments to CSS files that
documented which SASS file the code originated from. However, these comments use
full paths, so change depending on which computer was used to build the code:
this is not ideal when committing them to a Git repository.

This commit sets an appropriate option in the config.rb file to prevent these
comments from being generated, and removes the comments from the CSS files in
the build directory.

You can view the documentation for Compass' configuration here:
<http://compass-style.org/help/tutorials/configuration-reference/>

Change-Id: I0d312e93bb3615640ecc91462dfd6f96bde1bed7